acces admin bloqués  pour les users

+++ Workshop_web_s2/server.py
@@ -15,34 +15,6 @@ def index():
     sessions_per_category = model.get_number_of_sessions_per_category()
     infos_categories = model.combine_infos_categories(games_per_category, sessions_per_category)
     return render_template('home.html', infos_categories=infos_categories)
-# ---------------------------------------------------------------------------
-# ---------------------------- Admin
-# ---------------------------------------------------------------------------
-
-@app.route("/admin_games")
-def admin_games():
-    data = model.get_all_games()
-    return render_template('admin/games/admin_games.html', games=data)
-
-@app.route("/games/admin/<id>", methods=['GET']) 
-def admin_game(id):
-    data = model.get_game(int(id))
-    return render_template('admin/games/admin_game.html', game=data)
-
-@app.route("/admin_categories")
-def admin_categories():
-    data = model.get_all_category()
-    return render_template('admin/categories/admin_categories.html', categories=data)
-
-@app.route("/admin_sessions")
-def admin_sessions():
-    sessions = model.get_all_sessions()
-    return render_template('admin/sessions/admin_sessions.html', sessions=sessions)
-
-@app.route("/admin_users")
-def admin_users():
-    data = model.get_all_users()
-    return render_template('admin/users/admin_users.html', users = data)
 
 @app.route("/search", methods=['GET'])
 def search():
@@ -53,52 +25,91 @@ def search():
     return
 
 # ---------------------------------------------------------------------------
-# ---------------------------- Users
+# ---------------------------- Admin
 # ---------------------------------------------------------------------------
 
-@app.route("/users")
-def users():
-    data = model.get_all_users()
-    return render_template('user/users.html', users=data)
+@app.route("/admin_games")
+def admin_games():
+    if 'admin' in session and session['admin']==1 : 
+        data = model.get_all_games()
+        return render_template('admin/games/admin_games.html', games=data)
+    return redirect ("/")
+
+@app.route("/games/admin/<id>", methods=['GET']) 
+def admin_game(id):
+    if 'admin' in session and session['admin']==1 : 
+        data = model.get_game(int(id))
+        return render_template('admin/games/admin_game.html', game=data)
+    return redirect ("/")
+
+@app.route("/admin_categories")
+def admin_categories():
+    if 'admin' in session and session['admin']==1 : 
+        data = model.get_all_category()
+        return render_template('admin/categories/admin_categories.html', categories=data)
+    return redirect ("/")
+
+@app.route("/admin_sessions")
+def admin_sessions():
+    if 'admin' in session and session['admin']==1 : 
+        sessions = model.get_all_sessions()
+        return render_template('admin/sessions/admin_sessions.html', sessions=sessions)
+    return redirect ("/")
+
+@app.route("/admin_users")
+def admin_users():
+    if 'admin' in session and session['admin']==1 :
+        data = model.get_all_users()
+        return render_template('admin/users/admin_users.html', users = data)
+    return redirect ("/")
+
+# ---------------------------------------------------------------------------
+# ---------------------------- Users
+# ---------------------------------------------------------------------------
    
 @app.route("/users/<id>", methods=['GET']) 
 def user(id):
     data = model.get_user(int(id))
-    if data[0]['admin']==1 :
+    if 'id_user' in session :
+        if session['admin']==0 :
+            if session['id_user']==int(id):   
+                return render_template('profile.html', user=data)
+            return redirect ("/")
         return render_template('profile.html', user=data)
-    return render_template('profile.html', user=data)
+    return redirect ("/")
 
 @app.route("/user/edit/<id>", methods=['GET', 'POST'])
 def edit_user(id):
-    if request.method == 'POST' :
-        username = request.form['username'].strip()
-        email = request.form['email'].strip()
-        nationality = request.form['nationality'].strip()
-        password = request.form['password']
-        if password :
-            hashed_password = model.hash_psw(password)
-        else :
-            hashed_password = None
-        model.update_user(int(id), email, username, hashed_password, nationality)
-        data = model.get_user(int(id))
-        if data[0]['admin']==1 :
+    if 'admin' in session and session['admin']==1 :  
+        if request.method == 'POST' :
+            username = request.form['username'].strip()
+            email = request.form['email'].strip()
+            nationality = request.form['nationality'].strip()
+            password = request.form['password']
+            if password :
+                hashed_password = model.hash_psw(password)
+            else :
+                hashed_password = None
+            model.update_user(int(id), email, username, hashed_password, nationality)
+            data = model.get_user(int(id))
+            if data[0]['admin']==1 :
+                return render_template('profile.html', user=data)
             return render_template('profile.html', user=data)
-        return render_template('profile.html', user=data)
-    data = model.get_user(int(id))
-    return render_template('user/edit_user.html', user=data)
+        data = model.get_user(int(id))
+        return render_template('user/edit_user.html', user=data)
+    return redirect ("/")
+
 
 @app.route("/user/delete/<id>", methods=['GET', 'POST'])
 def delete_user(id):
-    model.delete_one_user(int(id))
-    return redirect("/admin_users", code=302)
+    if 'admin' in session and session['admin']==1 :  
+        model.delete_one_user(int(id))
+        return redirect("/admin_users", code=302)
+    return redirect ("/")
 
 # ---------------------------------------------------------------------------
 # ---------------------------- Connexion/Inscription/Comptes
 # ---------------------------------------------------------------------------
-
-@app.route("/account")
-def account():
-    return render_template('account.html')
 
 @app.route("/signup", methods=['GET','POST'])
 def signup():
@@ -166,54 +177,60 @@ def game(id):
 
 @app.route("/games/add", methods=['GET', 'POST'])
 def add_game():
-    if request.method == 'POST' :
-        name = request.form['name'].strip()
-        description = request.form['description'].strip()
-        released_date = request.form['released_date']
-        image = request.form['image'].strip()
-        ctgs = request.form.getlist('category[]')
-        ctgs = [int(c) for c in ctgs]
-        model.add_new_game(name, description, released_date, image)
-        id_new_game = model.get_latest_game_id()[0]
-        for ctg in ctgs:
-            model.add_liaison_gc(id_new_game,ctg)
-    ctgs = model.get_all_category()
-    return render_template('admin/games/form_game.html', all_ctgs=ctgs)
+    if 'admin' in session and session['admin']==1 :      
+        if request.method == 'POST' :
+            name = request.form['name'].strip()
+            description = request.form['description'].strip()
+            released_date = request.form['released_date']
+            image = request.form['image'].strip()
+            ctgs = request.form.getlist('category[]')
+            ctgs = [int(c) for c in ctgs]
+            model.add_new_game(name, description, released_date, image)
+            id_new_game = model.get_latest_game_id()[0]
+            for ctg in ctgs:
+                model.add_liaison_gc(id_new_game,ctg)
+        ctgs = model.get_all_category()
+        return render_template('admin/games/form_game.html', all_ctgs=ctgs)
+    return redirect ("/")
 
 @app.route("/games/edit/<id>", methods=['GET', 'POST'])
 def edit_game(id):
-    if request.method == 'POST' :
-        name = request.form['name'].strip()
-        description = request.form['description'].strip()
-        released_date = request.form['released_date']
-        image = request.form['image'].strip()
-        model.update_game(int(id), name, description, released_date, image)
+    if 'admin' in session and session['admin']==1 : 
+        if request.method == 'POST' :
+            name = request.form['name'].strip()
+            description = request.form['description'].strip()
+            released_date = request.form['released_date']
+            image = request.form['image'].strip()
+            model.update_game(int(id), name, description, released_date, image)
 
-        #gestion category
-        previous_ctgs = model.get_categories_from_game_id(int(id))
-        # previous_ctgs = [ctg['id_category'] for ctg in previous_ctgs]
-        new_ctgs = request.form.getlist('category[]')
-        new_ctgs = [int(c) for c in new_ctgs]
+            #gestion category
+            previous_ctgs = model.get_categories_from_game_id(int(id))
+            # previous_ctgs = [ctg['id_category'] for ctg in previous_ctgs]
+            new_ctgs = request.form.getlist('category[]')
+            new_ctgs = [int(c) for c in new_ctgs]
 
-        for previous_ctg in previous_ctgs:
-            if previous_ctg not in new_ctgs:
-                model.delete_liaison_gc(id, previous_ctg)
+            for previous_ctg in previous_ctgs:
+                if previous_ctg not in new_ctgs:
+                    model.delete_liaison_gc(id, previous_ctg)
 
-        for new_ctg in new_ctgs:
-            if new_ctg not in previous_ctgs:
-                model.add_liaison_gc(id, new_ctg)
+            for new_ctg in new_ctgs:
+                if new_ctg not in previous_ctgs:
+                    model.add_liaison_gc(id, new_ctg)
 
-    data = model.get_game(int(id))
-    ctgs = model.get_all_category()
-    game_ctgs = model.get_categories_from_game_id(int(id))
-    # game_ctgs = [ctg['id_category'] for ctg in game_ctgs]
-    return render_template('admin/games/form_game.html', game=data, all_ctgs=ctgs, game_ctgs=game_ctgs)
+        data = model.get_game(int(id))
+        ctgs = model.get_all_category()
+        game_ctgs = model.get_categories_from_game_id(int(id))
+        # game_ctgs = [ctg['id_category'] for ctg in game_ctgs]
+        return render_template('admin/games/form_game.html', game=data, all_ctgs=ctgs, game_ctgs=game_ctgs)
+    return redirect ("/")
 
 @app.route("/games/delete/<id>", methods=['GET', 'POST'])
 def delete_game(id):
-    model.delete_one_game(id)
-    print (id)
-    return redirect("/admin_games", code=302)
+    if 'admin' in session and session['admin']==1 : 
+        model.delete_one_game(id)
+        print (id)
+        return redirect("/admin_games", code=302)
+    return redirect ("/")
 
 @app.route("/get_session_from_game_category", methods=['GET'])
 def get_session_from_game_and_category():
@@ -236,27 +253,33 @@ def get_categories(idGame):
 
 @app.route("/category/add", methods=['GET', 'POST'])
 def add_category():
-    if request.method == 'POST':
-        label = request.form.get('label').strip()
-        description = request.form.get('description').strip()
-        model.add_new_category(label,description)
-    return render_template('admin/categories/form_category.html')
+    if 'admin' in session and session['admin']==1 : 
+        if request.method == 'POST':
+            label = request.form.get('label').strip()
+            description = request.form.get('description').strip()
+            model.add_new_category(label,description)
+        return render_template('admin/categories/form_category.html')
+    return redirect ("/")
 
 @app.route("/categories/edit/<id>", methods=['GET', 'POST'])
 def edit_category(id):
-    if request.method == 'POST':
-        label = request.form.get('label').strip()
-        description = request.form.get('description').strip()
-        model.update_category(id,label,description)
-        data = model.get_all_category()
-        return render_template('admin/categories/admin_categories.html', categories=data)
-    category = model.get_category_from_id(id)
-    return render_template('admin/categories/form_category.html', ctg=category)
+    if 'admin' in session and session['admin']==1 : 
+        if request.method == 'POST':
+            label = request.form.get('label').strip()
+            description = request.form.get('description').strip()
+            model.update_category(id,label,description)
+            data = model.get_all_category()
+            return render_template('admin/categories/admin_categories.html', categories=data)
+        category = model.get_category_from_id(id)
+        return render_template('admin/categories/form_category.html', ctg=category)
+    return redirect ("/")
 
 @app.route("/categories/delete/<id>")
 def delete_category(id):
-    model.delete_category_from_id(id)
-    return redirect("/admin_categories", code=302)
+    if 'admin' in session and session['admin']==1 : 
+        model.delete_category_from_id(id)
+        return redirect("/admin_categories", code=302)
+    return redirect ("/")
 
 
 # ---------------------------------------------------------------------------
